# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoSteps API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChallengeCategory = typeof ChallengeCategory[keyof typeof ChallengeCategory];


export const ChallengeCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  waste: 'waste',
  water: 'water',
  shopping: 'shopping',
} as const;

export type ChallengeDifficulty = typeof ChallengeDifficulty[keyof typeof ChallengeDifficulty];


export const ChallengeDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Challenge {
  id: number;
  title: string;
  description: string;
  xpReward: number;
  ecoScore?: number;
  category: ChallengeCategory;
  icon: string;
  difficulty: ChallengeDifficulty;
  completed: boolean;
  /** @nullable */
  streak?: number | null;
}

export interface ChallengeResult {
  xpEarned: number;
  newTotalXp: number;
  levelUp: boolean;
  /** @nullable */
  newLevel?: number | null;
  /** @nullable */
  newBadge: string | null;
}

export interface Profile {
  id: number;
  username: string;
  level: number;
  xp: number;
  xpToNextLevel: number;
  ecoScore: number;
  streak: number;
  rank: number;
  totalChallengesCompleted: number;
  /** @nullable */
  avatarUrl?: string | null;
  title?: string;
}

export interface EcoStats {
  co2Saved: number;
  waterSaved: number;
  plasticAvoided: number;
  treesEquivalent: number;
  challengesThisWeek: number;
  challengesThisMonth: number;
}

export type BadgeCategory = typeof BadgeCategory[keyof typeof BadgeCategory];


export const BadgeCategory = {
  streak: 'streak',
  challenge: 'challenge',
  eco_score: 'eco_score',
  special: 'special',
} as const;

export type BadgeRarity = typeof BadgeRarity[keyof typeof BadgeRarity];


export const BadgeRarity = {
  common: 'common',
  rare: 'rare',
  epic: 'epic',
  legendary: 'legendary',
} as const;

export interface Badge {
  id: number;
  name: string;
  description: string;
  icon: string;
  earned: boolean;
  /** @nullable */
  earnedAt?: string | null;
  category: BadgeCategory;
  rarity?: BadgeRarity;
}

export interface RecyclingItem {
  id: number;
  name: string;
  canRecycle: boolean;
  instructions: string;
  tip: string;
  icon?: string;
}

export interface RecyclingCategory {
  id: number;
  name: string;
  icon: string;
  color: string;
  description?: string;
  items: RecyclingItem[];
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  level: number;
  xp: number;
  ecoScore: number;
  streak: number;
  isCurrentUser?: boolean;
  /** @nullable */
  avatarUrl?: string | null;
}