# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EcoSteps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenges/{id}/complete:
    post:
      operationId: completeChallenge
      tags: [challenges]
      summary: Mark a challenge as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge completed, XP awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResult"
...
components:
  schemas:
...
    ChallengeResult:
      type: object
      required: [xpEarned, newTotalXp, levelUp, newBadge]
      properties:
        xpEarned:
          type: integer
        newTotalXp:
          type: integer
        levelUp:
          type: boolean
        newLevel:
          type: ["integer", "null"]
        newBadge:
          type: ["string", "null"]
...