# lib/api-zod/src/generated/types/challengeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoSteps API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeResult {
  xpEarned: number;
  newTotalXp: number;
  levelUp: boolean;
  /** @nullable */
  newLevel?: number | null;
  /** @nullable */
  newBadge: string | null;
}